+++ CHANGELOG.md
@@ -4,10 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-
 ## [Unreleased]
 
 ...
+
+## [0.9.7] - 2019-08-20
+
+## Added
+
+- Added method `IsValidDatabaseName` (and table/column variants) to `QuerySyntaxHelper`.  This allows testing strings without try/catch
+
+### Fixed
+
+- Tables with invalid names e.g. `[mytbl.lol]][.lol.lol]` are no longer returned by `DiscoveredDatabase.DiscoverTables` (previously a `RuntimeNameException` was thrown)
 
 ## [0.9.6] - 2019-08-09
 
@@ -60,7 +69,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed Drop table to work correctly with Views
 - Exists now works correctly for Views (previously it would return true if there was no view but a table with the same name)
 
-[Unreleased]: https://github.com/HicServices/FAnsiSql/compare/0.9.6...develop
+[Unreleased]: https://github.com/HicServices/FAnsiSql/compare/0.9.7...develop
+[0.9.7]: https://github.com/HicServices/FAnsiSql/compare/0.9.6...0.9.7
 [0.9.6]: https://github.com/HicServices/FAnsiSql/compare/0.9.5...0.9.6
 [0.9.5]: https://github.com/HicServices/FAnsiSql/compare/0.9.4...0.9.5
 [0.9.4]: https://github.com/HicServices/FAnsiSql/compare/0.9.3...0.9.4